Handle discrepancies between number of colors and number of highlights
If only one color is provided, color all highlights in the same color and print a message. If the number of colors is less than the number of highlights, print a warning and recycle the last color. If the number of colors isgreater than the number of highlights, print a warning and ignore the excessive colors.